Practice/Dev/Authentication and API/FASTAPI/requirements.txt
fastapi
psycopg2-binary
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings